+++ TTO_MAIN/src/com/group/tto/main/dao/impl/StoreDaoImpl.java
@@ -24,7 +24,7 @@ public class StoreDaoImpl extends BaseDao<Store> implements StoreDao {
 
   @Override
 	//根据商家id找到商家
-  public Store getStoreById(int sid) {
+  public Store getStoreById(Long sid) {
     CriteriaBuilder builder = this.getEntityManager().getCriteriaBuilder();
     CriteriaQuery<Store> query = builder.createQuery(Store.class);
     Root<Store> store = query.from(Store.class);
